broken shit

+++ calculate.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Particle.h"
 
-float distance(Particle& src, Particle& dst);
-float delta_x(Particle& src, Particle& dst);
-float delta_y(Particle& src, Particle& dst);
-float gravitational_force(Particle& p0, Particle& p1, float radius);
+double distance(Particle& src, Particle& dst);
+double delta_x(Particle& src, Particle& dst);
+double delta_y(Particle& src, Particle& dst);
+double gravitational_force(Particle& p0, Particle& p1, double radius);
